added test for STL, OBJ and OBJ2 loaders

+++ wrapper/src/main/kotlin/org/three/lights/AmbientLight.kt
@@ -4,5 +4,5 @@ package org.three.lights
 
 open external class AmbientLight(
         color: Int = definedExternally,
-        intensity: Float = definedExternally
+        intensity: Double = definedExternally
 ) : Light
